Added optional onSubmitSuccess prop
This prop callback is bit redundant, but there should be a more explicit
callback that does exactly what it says. The existing `onSubmit` callback
is a bit misleading with the `returnRejectedSubmitPromise` flag on. This
should resolve some confusion. Users that use `onSubmit` can switch to
`onSubmitSuccess` and `onSubmitFail` whenever they want. This is a
backwards compatible change.